requirements.txt
requests
edge-tts
ffmpeg-python
google-auth-oauthlib
google-api-python-client
pyyaml
schedule
Pillow
numpy
google-generativeai
